Updated default values and fixed issue with multiple records saving
When doing an add and calculating -> A POST operation is performed as
expected

However, after it saves and pressing "Edit Inputs" and updating the
values and submitting again, it does another POST and saves as a
separate record.

This was because we were not updating teh self.property objects id
value. So when doing a saveOrUpdateSyncronously() in analysis-tab-form
on submit, it was still seeing it did not have an id and thought it was
a new object. Fixed that by having spring mvc send the generated id back
and setting to object